validate: Add support for .ignore files
The translation file format is quite basic, and if a string is
translated the same in English and French, it's not possible to
differentiate it from an untranslated string.
This commit adds support for .ignore files which will allow to specify
when such strings should be ignored (ie assumed to be translated even if
they are the same as the English version).