i18nstats: Iterate over en files in compare_translations_packs
It currently iterates over i18n files, but then builds an en filename
from the file being iterated. This reverses this logic as it feels more
natural this way (iterate over en files, get the corresponding i18n
file). This is made possible with the refactorings in the previous
commits.